requirements.txt
fastapi
uvicorn[standard]
python-multipart
PyPDF2
nltk
python-dotenv
torch
transformers
mistralai